pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "pybind11>=2.10"]
build-backend = "setuptools.build_meta"

[project]
name = "autolife_planning"
version = "0.1.1"
description = "Autolife planning module"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy",
    "scipy",
    "pybind11>=2.10",
    "autolife-vamp>=0.6.1",
]

[project.optional-dependencies]
dev = ["pre-commit", "pybullet"]
pinocchio = ["pin"]

[tool.setuptools.packages.find]
include = ["autolife_planning*"]

[tool.setuptools.package-data]
autolife_planning = [
    "resources/**/*.urdf",
    "resources/**/*.srdf",
    "resources/**/*.STL",
    "resources/**/*.ply",
]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.pypi-dependencies]
autolife_planning = { path = ".", editable = true }
autolife-vamp = { path = "./third_party/vamp", editable = true }
foam = { path = "./third_party/foam", editable = true }

[tool.pixi.environments]
default = { solve-group = "default" }
dev = { features = ["dev"], solve-group = "default" }

[tool.pixi.tasks]
cricket-configure = "cmake -GNinja -S third_party/cricket -B third_party/cricket/build"
cricket-build = { cmd = "cmake --build third_party/cricket/build", depends-on = ["cricket-configure"] }
foam-configure = "cmake -GNinja -S third_party/foam -B third_party/foam/build"
foam-build = { cmd = "cmake --build third_party/foam/build", depends-on = ["foam-configure"] }
build-robot = { cmd = "bash scripts/build_robot.sh", depends-on = ["foam-build"] }
spherize-robot = { cmd = "bash scripts/spherize_robot.sh", depends-on = ["build-robot"] }
generate-fk = { cmd = "bash scripts/generate_fk.sh", depends-on = ["spherize-robot"] }
build-vamp = "python -m build third_party/vamp --no-isolation --skip-dependency-check --outdir dist/"
build-pkg = "python -m build --no-isolation --skip-dependency-check --outdir dist/"
build-all = { depends-on = ["build-vamp", "build-pkg"] }

[tool.pixi.dependencies]
cmake = ">=4.2.3,<5"
ninja = ">=1.13.2,<2"
cxx-compiler = ">=1.11.0,<2"
pkg-config = ">=0.29.2,<0.30"
git = ">=2.53.0,<3"
patch = ">=2.8,<3"
python = ">=3.11,<3.13"
pybind11 = ">=3.0.2,<4"
libboost-devel = "*"
orocos-kdl = ">=1.5.3,<2"
eigen = ">=3.4.0,<4"
nlopt = ">=2.10.0,<3"
pinocchio = ">=3.9.0,<4"
cppad = ">=20260000.0,<20260001"
cgal = ">=6.0.1,<7"
inja = ">=3.5.0,<4"
nlohmann_json = ">=3.12.0,<4"
fmt = ">=12.1.0,<13"
llvm = ">=22.1.0,<23"
fire = ">=0.7.1,<0.8"
xmltodict = ">=1.0.4,<2"
pymeshlab = ">=2025.7.post1,<2026"
pip = ">=26.0.1,<27"
python-build = ">=1.2.2,<2"
scikit-build-core = ">=0.4.3"
nanobind = ">=2.0.0"

[tool.pixi.feature.dev.pypi-dependencies]
pybullet = "*"
